refactor: folder restructure and import clean up
