+++ src/components/gallery/gallery-item.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Image from 'next/image';
-import { GalleryItemData } from '@/content/gallery';
+import { GalleryItemData } from '@content/gallery';
 
 interface GalleryItemProps {
     galleryItemData: GalleryItemData;
@@ -15,7 +15,7 @@ const GalleryItem = ({ galleryItemData, selectGalleryItem, isLarge }: GalleryIte
         <div className={itemClass} onClick={() => selectGalleryItem(galleryItemData.id)}>
             <Image
                 style={{ objectFit: 'cover' }}
-                src={require(`../../public/images/thumbs/${galleryItemData.src}`)}
+                src={require(`../../../public/images/thumbs/${galleryItemData.src}`)}
                 alt={`Thumbnail - ${galleryItemData.name}`}
                 fill
                 sizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
