+++ src/components/shop/shop-item.tsx
